app/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-filter==22.1
#django-filter==2.0.0
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2-binary==2.9.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
fontawesomefree==6.1.2

